requirements.txt
numpy==2.2.3
pandas==2.2.3
matplotlib==3.10.1
scipy==1.12.0
scikit-learn==1.4.2
tensorflow==2.15.0
keras==2.15.0
torch==2.1.2
xgboost==2.0.3
PyQt5==5.15.10
opencv-python==4.8.1.78
pytesseract==0.3.10
joblib==1.3.2
scapy==2.5.0
pyshark==0.6.0
pyusb==1.2.1
pyserial==3.5
plotly==6.0.0
imutils==0.5.4
pillow==11.1.0
h5py==3.10.0
tqdm==4.67.1
statsmodels==0.14.1
gym==0.26.0
stable-baselines3==2.2.1
nltk==3.8.1
matplotlib-inline==0.1.6
seaborn==0.13.1
ipykernel==6.29.0
jupyter==1.0.0
black==24.1.1
flake8==7.0.0
pytest==8.0.0
pyarrow==19.0.1
dask==2023.3.0
ray==2.9.0
psutil==5.9.8
imageio==2.33.1
sympy==1.12.1
cython==3.0.7
pybluez==0.23.0